+++ SPUIView/View/Result/Controller/SVDetailViewCtrl.h
@@ -13,7 +13,7 @@
 #import <UIKit/UIKit.h>
 
 @interface SVDetailViewCtrl : UIViewController
-@property long testId; // 测试ID
+@property long long testId; // 测试ID
 @property NSString *testType; // 测试类型：0=video,1=web,2=speed
 
 @end
